user commit

+++ app/Http/Controllers/PostController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+
 use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Routing\Controller as BaseController;
 use Illuminate\Foundation\Validation\ValidatesRequests;
@@ -9,8 +10,15 @@ use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use App\Post;
 use App\Category;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+
 class PostController extends BaseController
 {
+
+    public function _construct(){
+        $this->middleware('auth',['only'=>'index', 'create', 'store', 'update']);
+    }
     public function index(){
         $posts=Post::orderBy('created_at', 'desc')->paginate(4); //gaunu duom
         $categories=Category::all(); //gaunu duom
@@ -23,28 +31,36 @@ class PostController extends BaseController
 
 
 
+
     public function create( ){
+
+        $user=Auth::user();
         $categories=Category::all();
         return view('pages.forma')
-            ->with("categories",$categories);
+            ->with("categories",$categories)
+            ->with('user', $user);
     }
     public function store(Request $request){
      $request->validate([
          'title'=>'required',
          'description'=>'required',
          'img'=>'image|nullable|max:2048',
+         'cat_id'=>'required',
+         'user_id'=>'required'
      ]);
 
         $imagePath="";
      if ($request['img'] ){
          $imagePath = $request->file('img')->store('public/images');
      }
+     $imagePath2=str_replace("public", "storage", $imagePath);
 
       Post::create([
          'title'=>request('title'),
          'description'=>request('description'),
-         'img'=>$imagePath,
-         'cat_id'=>request('cat_id')
+         'img'=>$imagePath2,
+         'cat_id'=>request('cat_id'),
+         'user_id'=>request('user_id')
 
      ]);
 
@@ -75,8 +91,13 @@ class PostController extends BaseController
 
     public function edit($id)
     {
-        $categories=Category::all();
         $post = Post::find($id);
+        $categories=Category::all();
+        if(Gate::denies('edit-post', $post)){
+            return view('auth.login');
+        }
+
+
         return view('posts.edit-post')
             ->with('post', $post)
             ->with("categories",$categories);
@@ -93,7 +114,7 @@ class PostController extends BaseController
             'img' => 'image|nullable|max:1999'
         ]);
 
-  //      if($request->hasFile('img')){
+//        if($request->hasFile('img')){
 //           File::delete('storage/app/'.$post->img);
 //
 //            $filenameWithExt = $request->file('image')->getClientOriginalName(); // Full file name
@@ -101,9 +122,9 @@ class PostController extends BaseController
 //            $extension = $request->file('image')->guessClientExtension(); //Only file extesion name
 //            $fileName = $filename.'_'.time().'.'.$extension; //Concatenated file name and extension with custom naming extensions
 //       }
-
-        // Image path ready to upload
-        //$request->file('post-image')->storeAs('public/posts-images', $fileName);
+//
+//         //Image path ready to upload
+//        $request->file('post-image')->storeAs('public/posts-images', $fileName);
 
         Post::where('id',$id)->update([
             'title' => request('title'),
@@ -119,9 +140,13 @@ class PostController extends BaseController
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(Gate::denies('delete-post', $post)){
+            return view('auth.login');
+        }
+            //File::delete('storage/app/'.$post->img);
+            $post->delete();
 
-        //File::delete('storage/app/'.$post->img);
-        $post->delete();
+
 
         return redirect('/');
     }
